move retrier to separate package and reuse code
